+++ packages/core/src/aminomessages/posts/index.ts
@@ -1,3 +1,4 @@
 export * from "./converter";
 export * from "./messages";
 export * from "./registry";
+export * from "./types";
