+++ packages/core/src/aminomessages/posts/messages.ts
@@ -2,34 +2,10 @@ import { AminoMsg } from "@cosmjs/amino";
 import Long from "long";
 import {
   Entities,
-  Poll_ProvidedAnswer,
-  PollTallyResults,
   PostReference,
   ReplySetting,
 } from "@desmoslabs/desmjs-types/desmos/posts/v1/models";
-import { EncodeObject } from "@cosmjs/proto-signing";
-
-export interface AminoAttachment extends AminoMsg {}
-
-export interface AminoPoll extends AminoAttachment {
-  readonly type: "desmos/Poll";
-  readonly value: {
-    question: string;
-    provided_answers: Poll_ProvidedAnswer[];
-    end_date?: Date;
-    allows_multiple_answers: boolean;
-    allows_answer_edits: boolean;
-    final_tally_results?: PollTallyResults;
-  };
-}
-
-export interface AminoMedia extends AminoAttachment {
-  readonly type: "desmos/Media";
-  readonly value: {
-    uri: string;
-    mime_type: string;
-  };
-}
+import { AminoAttachment } from "./types";
 
 export interface AminoMsgCreatePost extends AminoMsg {
   readonly type: "desmos/MsgCreatePost";
